Moved interfaces description in separated file
